+++ app/src/main/java/com/vaudibert/canidrive/ui/DriveFragment.kt
@@ -2,16 +2,27 @@ package com.vaudibert.canidrive.ui
 
 
 import android.os.Bundle
+import android.os.Handler
+import android.os.Looper
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.vaudibert.canidrive.R
+import com.vaudibert.canidrive.domain.Drinker
+import kotlinx.android.synthetic.main.fragment_drive.*
+import java.util.*
 
 /**
  * A simple [Fragment] subclass.
  */
 class DriveFragment : Fragment() {
+
+    lateinit var drinker: Drinker
+
+    lateinit var mainHandler: Handler
+
+    lateinit var pastDrinksAdapter: PastDrinksAdapter
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -21,5 +32,51 @@ class DriveFragment : Fragment() {
         return inflater.inflate(R.layout.fragment_drive, container, false)
     }
 
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
 
+        drinker = (this.activity as MainActivity).drinker
+
+        pastDrinksAdapter =
+            PastDrinksAdapter(
+                this.context!!,
+                drinker.getDrinks()
+            )
+        listViewPastDrinks.adapter = pastDrinksAdapter
+
+        // needed for periodic update of drinker status
+        mainHandler = Handler(Looper.getMainLooper())
+
+
+    }
+    private fun updateDriveStatus() {
+        textViewDriveStatus.text = if (drinker.alcoholRateAt(Date()) < 0.5)
+            "DRIVE : YES"
+        else
+            "DRIVE : NO"
+
+        val drinks = drinker.getDrinks()
+        listViewPastDrinks.adapter =
+            PastDrinksAdapter(this.context!!, drinks)
+    }
+
+    /**
+     * Helper task to update the drive status while app is running.
+     */
+    private val updateDriveStatusTask = object : Runnable {
+        override fun run() {
+            updateDriveStatus()
+            mainHandler.postDelayed(this, 1000*60)
+        }
+    }
+
+    override fun onPause() {
+        super.onPause()
+        mainHandler.removeCallbacks(updateDriveStatusTask)
+    }
+
+    override fun onResume() {
+        super.onResume()
+        mainHandler.post(updateDriveStatusTask)
+    }
 }
